Add preconditioned Crank-Nicholson algtype for MCMC
Add support for a new `MCMC.algtype`, `"pCN"` (in addition to previously
implemented `"rwm"` for random-walk Metropolis) in MCMC proposal() and
mcmc_sample!() steps. Update docstrings.

Refactor existing MCMC unit tests

Add unit test for pCN method in MCMC

Change random seed to fix unit test

Stepsize test for pCN (l.171) failed on Julia 1.5.4:
https://github.com/CliMA/CalibrateEmulateSample.jl/actions/runs/1551172081
verified it failed on 1.6 but passes on 1.7. This indicates it's due to
the global RNG change in 1.7 (acceptance ratio was close to one of the
thresholds used in the logic). Left logic unchanged but selected a new
RNG value for which tests pass on both 1.6 and 1.7.

Fix cov scaling w/stepsize, & typos