Specify RNG algorithm in unit tests
In all unit tests involving randomness, create an `rng` object and
explicitly pass it to all samplers. This will simplify maintaining
tests when Julia's RNG implementation changes, as it did in 1.6 -> 1.7.

NOTE: This introduces a new error in unit tests for Utilities.jl, as it
calls the EnsembleKalmanProcess constructor with an `rng` kwarg, which
was intoduced in
tsj5/EnsembleKalmanProcesses.jl@d7b735c03fb5fc397d54173f856650f6f39dc5bc
This error will be resolved when/if the associated PR
( CliMA/EnsembleKalmanProcesses.jl#81 ) is merged.